+++ extensions/Worker.Extensions.Storage.Blobs/src/BlobInputAttribute.cs
@@ -1,15 +1,16 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using Microsoft.Azure.Functions.Worker.Converters;
 using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
 
 namespace Microsoft.Azure.Functions.Worker
 {
+    [InputConverter(typeof(BlobStorageConverter))]
+    [ConverterFallbackBehavior(ConverterFallbackBehavior.Default)]
     public sealed class BlobInputAttribute : InputBindingAttribute
     {
         private readonly string _blobPath;
-
-        private bool _isBatched = false;
 
         /// <summary>Initializes a new instance of the <see cref="BlobInputAttribute"/> class.</summary>
         /// <param name="blobPath">The path of the blob to which to bind.</param>
@@ -24,16 +25,6 @@ namespace Microsoft.Azure.Functions.Worker
         public string BlobPath
         {
             get { return _blobPath; }
-        }
-
-        /// <summary>
-        /// Gets or sets the configuration to enable batch processing of blobs. Default value is "false".
-        /// </summary>
-        [DefaultValue(false)]
-        public bool IsBatched
-        {
-            get => _isBatched;
-            set => _isBatched = value;
         }
 
         /// <summary>