+++ extensions/Worker.Extensions.Storage.Blobs/src/Constants.cs
@@ -1,15 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace Microsoft.Azure.Functions.Worker
+namespace Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs
 {
     internal static class Constants
     {
         internal const string Storage = "Storage";
         internal const string BlobExtensionName = "AzureStorageBlobs";
-        internal const string Connection = "Connection";
-        internal const string ContainerName = "ContainerName";
-        internal const string BlobName = "BlobName";
 
         // Media content types
         internal const string JsonContentType = "application/json";