+++ extensions/Worker.Extensions.Abstractions/release_notes.md
@@ -6,4 +6,4 @@
 
 ### Microsoft.Azure.Functions.Worker.Extensions.Abstractions <version>
 
-- <event>
+- <entry>