+++ sdk/release_notes.md
@@ -4,13 +4,15 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Sdk <version> (meta package)
+### Microsoft.Azure.Functions.Worker.Sdk <version>
 
 - <entry>
 
-### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version> (delete if not updated)
+### Microsoft.Azure.Functions.Worker.Sdk.Analyzers 1.2.0
 
-- <entry>
+- Add analyzer for SupportsDeferredBindingAttribute (#1367)
+- Added an analyzer that will show a warning for types not supported by a binding attribute (#1505)
+- Added an analyzer that will suggest a code refactor for all of the types supported by a binding attribute (#1604)
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>